+++ packages/navigation/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @d4kmor/navigation
 
+## 0.1.2
+
+### Patch Changes
+
+- Improve sidebar navigation by always linking to first sub page item and indicationg currently active anchor page
+
 ## 0.1.1
 
 ### Patch Changes